Add QUndoStack to splines
This is not very modular since spline is just a mathematical concept,
but it is far more convenient than separated undo stack.
QUndoStack can not be copied, so it is allocated dynamically and must have
a parent. Stack is not a Q class, so some Q parent is needed. But why
parent can be NULL? Because QHash demands its values having default
constructor, so NULL-check appeared to avoid dangling pointers to stacks.