Implement basic spline creation and viewing
Canvas::radius and Spline::approx are just adjustable params.
Spline::extSupport is not really a support, it should be renamed.
Checks like (spline->size() > 0) appear in for loops because
size() is unsigned, so size - 1 is a huge positive number even
when size == 0. When loop accepts it, it is logically incorrect.